Homework5
Code to distinguish between an input of 2 or 4 positive integers and declare the largest number. Splits the input into an array that is then stored for use with the LargestNumber method to return the largest number. Once output is displayed for largest integer, a call to method CreateAccount is conducted to request information for account creation while ensuring both passwords == and utilizing DateTime.Now.Year - input birth year to verify the person creating the account is above the age of 18.

Changes to consider for real world applications:
o an input of negative integer creates an error due to Convert.ToInt32 not having the ability to identify null integers, could correct with parsing which would require a slight redesign of current code.
o an input of year only is a limited assurance of age. Could implement mm/dd/yyyy input and date checking to provide a more definitive result.